// apps/mobile/lib/location-types.ts
export enum SportType {
  Basketball = "Basketball",
  Soccer = "Soccer",
  Badminton = "Badminton",
  Baseball = "Baseball",
  Volleyball = "Volleyball",
  Tennis = "Tennis",
  Pickleball = "Pickleball",
  American_Football = "American Football"
}